+++ api/lib/infrastructure/serializers/jsonapi/prescriber-serializer.js
@@ -24,7 +24,7 @@ module.exports = {
         'firstName',
         'lastName',
         'pixOrgaTermsOfServiceAccepted',
-        'areNewYearSchoolingRegistrationsImported',
+        'areNewYearOrganizationLearnersImported',
         'lang',
         'memberships',
         'userOrgaSettings',
