+++ api/lib/infrastructure/serializers/jsonapi/sup-organization-learner-warnings-serializer.js
@@ -2,7 +2,7 @@ const { Serializer } = require('jsonapi-serializer');
 
 module.exports = {
   serialize(organizations, meta) {
-    return new Serializer('higher-schooling-registration-warnings', {
+    return new Serializer('sup-organization-learner-warnings', {
       attributes: ['warnings'],
       meta,
     }).serialize(organizations);
